refactor(cni): promote hostgw out of the veth plugin's tree
Three shared packages moved out from under internal/cni in #315 so the
other chain binaries stop reaching into one plugin's own tree for things
all of them use. hostgw met the same description and stayed behind.

It is the strongest case of the four, not the weakest: hostconf,
crdnames and nadpatch are schema and naming helpers, while hostgw
configures kernel state on an interface, and two binaries outside its
parent (galactic-tap, via internal/cnitap; galactic-veth's own
internal/cni) call it to do that.

The move is mechanical. The package name does not change, every call
site is an import line, and no behavior changes.

While updating the architecture reference, its module tree still showed
hostconf, crdnames and nadpatch nested under internal/cni as well. That
has been stale since #315 promoted them; the tree and the unit-test
package list now match the filesystem for all four.

Related to #344
